pyproject.toml
[tool.poetry]
name = "ecommerce-pro"
version = "0.1.0"
description = ""
authors = ["Jose Felipe <felipe.oliveira@blitzpay.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
python-decouple = "^3.8"
dj-database-url = "^2.1.0"
psycopg2 = "^2.9.9"
pillow = "^10.1.0"
django-admin-thumbnails = "^0.2.8"
mercadopago = "^2.2.1"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
